requirements.txt
fastapi
uvicorn
pydantic
gitpython

